personalized ats score checker_GenAI model/requirement.txt
streamlit
google.generativeai
os
PyPDF2
dotenv
json